Adds function and tests for interative and constant game simulation.
I still need a way to match the individual players into game pairs
while minimizing the number of terminating games. I'll probably just
make a grid and check the sum of the rows, then return the min. It's
an n^2 operation but the test only gives you 1 to 100 players so
hopefully it's a managable task.